+++ R/EngineRestGitHub.R
@@ -61,9 +61,9 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
     #' @param settings A list of  `GitStats` settings.
     #' @return A table of commits.
     pull_commits = function(org,
-                           date_from,
-                           date_until = Sys.date(),
-                           settings) {
+                            date_from,
+                            date_until = Sys.date(),
+                            settings) {
       NULL
     },
 
@@ -116,7 +116,7 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
         if (!private$scan_all) {
           cli::cli_alert_info("[GitHub][Engine:{cli::col_green('REST')}][org:{unique(repos_table$organization)}] Pulling contributors...")
         }
-        repo_iterator <- paste0(repos_table$organization, "/", repos_table$name)
+        repo_iterator <- paste0(repos_table$organization, "/", repos_table$repo_name)
         user_name <- rlang::expr(.$login)
         repos_table$contributors <- purrr::map_chr(repo_iterator, function(repos_id) {
           contributors_endpoint <- paste0(self$rest_api_url, "/repos/", repos_id, "/contributors")
@@ -229,31 +229,25 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
                                byte_max) {
       if (total_n >= 0 & total_n < 1e3) {
         resp_list <- list()
-
         for (page in 1:(total_n %/% 100)) {
           resp_list <- self$response(
             paste0(search_endpoint, "+size:0..", byte_max, "&page=", page, "&per_page=100")
           )[["items"]] %>%
             append(resp_list, .)
         }
-
         resp_list
       } else if (total_n >= 1e3) {
         resp_list <- list()
         index <- c(0, 50)
-
         spinner <- cli::make_spinner(
           which = "timeTravel",
           template = cli::col_grey(
             "GitHub search limit (1000 results) exceeded. Results will be divided. {spin}"
           )
         )
-
         while (index[2] < as.numeric(byte_max)) {
           size_formula <- paste0("+size:", as.character(index[1]), "..", as.character(index[2]))
-
           spinner$spin()
-
           n_count <- tryCatch(
             {
               self$response(paste0(search_endpoint, size_formula))[["total_count"]]
@@ -262,7 +256,6 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
               NULL
             }
           )
-
           if (is.null(n_count)) {
             NULL
           } else if ((n_count - 1) %/% 100 > 0) {
@@ -273,9 +266,7 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
             resp_list <- self$response(paste0(search_endpoint, size_formula, "&page=1&per_page=100"))[["items"]] %>%
               append(resp_list, .)
           }
-
           index[1] <- index[2]
-
           if (index[2] < 1e3) {
             index[2] <- index[2] + 50
           }
@@ -300,8 +291,8 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
     tailor_repos_info = function(repos_list) {
       repos_list <- purrr::map(repos_list, function(repo) {
         list(
-          "id" = repo$id,
-          "name" = repo$name,
+          "repo_id" = repo$id,
+          "repo_name" = repo$name,
           "default_branch" = repo$default_branch,
           "stars" = repo$stargazers_count,
           "forks" = repo$forks_count,
@@ -322,7 +313,7 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
     # @return A table of repositories with added information on issues.
     pull_repos_issues = function(repos_table) {
       if (nrow(repos_table) > 0) {
-        repos_iterator <- paste0(repos_table$organization, "/", repos_table$name)
+        repos_iterator <- paste0(repos_table$organization, "/", repos_table$repo_name)
         issues <- purrr::map_dfr(repos_iterator, function(repo_path) {
           issues_endpoint <- paste0(self$rest_api_url, "/repos/", repo_path, "/issues")
 
@@ -363,8 +354,8 @@ EngineRestGitHub <- R6::R6Class("EngineRestGitHub",
     pull_commits_from_org = function(repos_table,
                                      date_from,
                                      date_until) {
-      repos_names <- repos_table$name
-      repo_fullnames <- paste0(repos_table$organization, "/", repos_table$name)
+      repos_names <- repos_table$repo_name
+      repo_fullnames <- paste0(repos_table$organization, "/", repos_table$repo_name)
 
       repos_list_with_commits <- purrr::map(repo_fullnames, function(repo_fullname) {
         commits_from_repo <- private$pull_commits_from_repo(