+++ R/GitHost.R
@@ -385,23 +385,16 @@ GitHost <- R6::R6Class("GitHost",
 
     # @description Add `api_url` column to table.
     add_repo_api_url = function(repos_table){
-      if ("file_content" %in% colnames(repos_table)) {
-        repo_name <- rlang::expr("repository_name")
-        repo_id <- rlang::expr("repository_id")
-      } else {
-        repo_name <- rlang::expr("name")
-        repo_id <- rlang::expr("id")
-      }
       if (!is.null(repos_table) && nrow(repos_table) > 0) {
         repos_table <- if (private$host == "GitHub") {
           dplyr::mutate(
             repos_table,
-            api_url = paste0(private$api_url, "/repos/", organization, "/", eval(parse(text = repo_name))),
+            api_url = paste0(private$api_url, "/repos/", organization, "/", repo_name),
           )
         } else if (private$host == "GitLab") {
           dplyr::mutate(
             repos_table,
-            api_url = paste0(private$api_url, "/projects/", gsub("gid://gitlab/Project/", "", eval(parse(text = repo_id))))
+            api_url = paste0(private$api_url, "/projects/", stringr::str_match(repo_id, "[0-9].*"))
           )
         }
       }